Changed game from 10FPS to 50FPS
This was done by changing the movement speed of the player and the game tickspeed that the game was playing at. However, when initially designing the game, the wall hit detection was designed for the player to move 10 pixels at a time. Since the player now moves 2 pixels at a time, this causes some clipping through the walls.